requirements.txt
annotated-types==0.7.0
blinker==1.8.2
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
numpy==2.0.2
packaging==24.1
pandas==2.2.3
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
